+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release Notes
 
-## 1.0.0
+## 0.1.0
 
 Initial Release
+
+## 0.2.0
+
+This is the minor release with some fixes in docs and addition of new functions.
+
+- Add `getDistrictsWithMunicipalities()` to district entity.
+- Add `getProvincesWithDistrictsWithMunicipalities()` to province entity.